requirements.txt
Python 3.13.8
pygame-ce 2.5.5